+++ packages/adapter-webpack/test/fixtures/react/adapter_options.js
@@ -1,6 +1,7 @@
 const { resolve } = require('path')
 
 const [clientConfig, serverConfig] = require('./webpack.config')
+const { filesForComponent, filesForVariant } = require('./scaffolding')
 const serverRenderPath = resolve(__dirname, 'server-render.js')
 const clientRenderPath = resolve(__dirname, 'client-render.js')
 const extensions = ['js', 'jsx']
@@ -12,5 +13,7 @@ module.exports = {
   serverRenderPath,
   clientRenderPath,
   extensions,
-  properties
+  properties,
+  filesForComponent,
+  filesForVariant
 }