+++ test/project/lib/webpack-adapter-react-options.js
@@ -1,14 +1,19 @@
 const { resolve } = require('path')
 
 const [clientConfig, serverConfig] = require('../webpack/react.conf')
+const { filesForComponent, filesForVariant } = require('./react-scaffolding')
 const serverRenderPath = resolve(__dirname, 'react-server-render.js')
 const clientRenderPath = resolve(__dirname, 'react-client-render.js')
 const extensions = ['js', 'jsx']
+const properties = 'prop-types'
 
 module.exports = {
   serverConfig,
   clientConfig,
   serverRenderPath,
   clientRenderPath,
-  extensions
+  extensions,
+  properties,
+  filesForComponent,
+  filesForVariant
 }